Move schema navigation to a separate component
