+++ src/vs/workbench/contrib/positronConnections/browser/components/schemaNavigation.css
@@ -1,0 +1,83 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2022 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+.connections-items-container {
+	height: 100%;
+	position: relative;
+	left: 0.5px;
+}
+
+.connections-item {
+	display: flex;
+	cursor: pointer;
+	align-items: center;
+	height: 26px;
+}
+
+.connections-details {
+	flex-grow: 1;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	font-family: monospace;
+	display: flex;
+	align-items: center;
+}
+
+.connections-dtype {
+	color: var(--vscode-positronSideActionBar-disabledForeground);
+	font-weight: 300;
+}
+
+.connections-dtype::before {
+	content: ":";
+	color: var(--vscode-positronVariables-foreground);
+	margin-left: 3px;
+	margin-right: 3px;
+}
+
+.connections-language::before {
+	content: ":";
+	color: var(--vscode-positronVariables-foreground);
+	margin-left: 0px;
+	margin-right: 6px;
+}
+
+.connections-icon {
+	flex-shrink: 0;
+	flex-grow: 0;
+	padding-right: 5px;
+	padding-left: 5px;
+}
+
+.connections-icon.disabled {
+	opacity: 0.7;
+}
+
+.connections-item .expand-collapse-area {
+	width: 26px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.connections-item.selected {
+	color: var(--vscode-positronVariables-activeSelectionForeground);
+	background: var(--vscode-positronVariables-activeSelectionBackground);
+	outline: 1px;
+	outline-offset: -1px;
+}
+
+.connection-disabled {
+	color: var(--vscode-positronSideActionBar-disabledForeground);
+}
+
+.connections-error {
+	margin-left: 5px;
+}
+
+.connections-error.codicon {
+	color: var(--vscode-debugConsole-errorForeground);
+}
