+++ src/vs/workbench/contrib/positronConnections/browser/positronConnectionsContext.tsx
@@ -1,0 +1,39 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import React, { PropsWithChildren, createContext, useContext } from 'react';
+import { IReactComponentContainer } from 'vs/base/browser/positronReactRenderer';
+import { ICommandService } from 'vs/platform/commands/common/commands';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
+import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
+import { IHoverService } from 'vs/platform/hover/browser/hover';
+import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
+import { IPositronConnectionsService } from 'vs/workbench/services/positronConnections/browser/interfaces/positronConnectionsService';
+
+export interface PositronConnectionsServices {
+	readonly commandService: ICommandService;
+	readonly configurationService: IConfigurationService;
+	readonly contextKeyService: IContextKeyService;
+	readonly contextMenuService: IContextMenuService;
+	readonly hoverService: IHoverService;
+	readonly keybindingService: IKeybindingService;
+	readonly connectionsService: IPositronConnectionsService;
+	readonly reactComponentContainer: IReactComponentContainer;
+}
+
+const PositronConnectionsContext = createContext<PositronConnectionsServices>(undefined!);
+
+export const PositronConnectionsContextProvider = (
+	props: PropsWithChildren<PositronConnectionsServices>
+) => {
+	return (
+		<PositronConnectionsContext.Provider value={props}>
+			{props.children}
+		</PositronConnectionsContext.Provider>
+	);
+};
+
+export const usePositronConnectionsContext = () => useContext(PositronConnectionsContext);
